Move costraint check to breeze
We started to copy more and more code from breeze utils to the
constraint generation script - including using github token for
pulling constraints files, reading python version, running command
etc - so it's now the time to move the constraint check to become
a regular breeze command and import utils from breeze utils.

We are running the command in the CI image which has both
advantages and disadvantages - it's a bit slower and requires
the CI image to be built for the right Python version but it
makes sure that the outputs are reproducible across different
OS's and system libraries.
