+++ dev/breeze/doc/images/output_release-management.txt
@@ -1,1 +1,1 @@
-ff286e72725eee4f8f720f4b95e14839
+3c496362e973303c4a018cf8db678482
