+++ dev/breeze/doc/images/output_setup_regenerate-command-images.txt
@@ -1,1 +1,1 @@
-d112f32bf3ac63a45cafb3cd1248d728
+b6bec540dff082eb353c54319297ed94
